requirements.txt
python-telegram-bot
requests
gitpython
